refactor(webview-ext): split framework-agnostic webview entry from react entry
